Added audio component to multiplayer mode. Fixed bug where controller's on client side would be in incorrect position due to position, rotation and scale not being set over data transfer, had to add a decomposition of the matrix sent over the connection for the client's entity to allow for CONTROLLER updates to be correct. I think using two sets of those vectors over complicates it and if someone can come up with a easier soltuion it might make the process cleaner. Also if the init/initComplexShape functions could be done on the physics constructor then it would avoid having to do it in the client file which is messy. Added to the renderer if a model isActive is false then it does not render.

+++ Engine/Network/Server/GameServer.hpp
@@ -45,8 +45,8 @@ namespace Engine
 		// Handle a request for entity data
 		void HandleRequestEntityData(int);
 
-		// Reset an entity and its component used by a disconnected client
-		void ResetNetworkEntity(uint64_t);
+		// Reset an entity and its component used by a disconnected client index
+		void ResetNetworkEntity(int);
 
 		// Clean up server memory using yojimbo
 		void CleanUp();
